Breakout compile functionality into a package.
