+++ nucypher/blockchain/eth/sol/compile/exceptions.py
@@ -1,0 +1,27 @@
+"""
+This file is part of nucypher.
+
+nucypher is free software: you can redistribute it and/or modify
+it under the terms of the GNU Affero General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+nucypher is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU Affero General Public License for more details.
+
+You should have received a copy of the GNU Affero General Public License
+along with nucypher.  If not, see <https://www.gnu.org/licenses/>.
+"""
+
+
+class CompilationError(RuntimeError):
+    """
+    Raised when there is a problem compiling nucypher contracts
+    or with the expected compiler configuration.
+    """
+
+
+class ProgrammingError(RuntimeError):
+    """Caused by a human error in code"""
