+++ nucypher/blockchain/eth/sol/source/contracts/NuCypherToken.sol
@@ -8,7 +8,7 @@ import "zeppelin/token/ERC20/ERC20Detailed.sol";
 
 
 /**
-* @title NuCypher token
+* @title NuCypherToken
 * @notice ERC20 token
 * @dev Optional approveAndCall() functionality to notify a contract if an approve() has occurred.
 */
