+++ nucypher/blockchain/eth/sol/source/contracts/StakingEscrow.sol
@@ -39,6 +39,7 @@ interface WorkLockInterface {
 
 
 /**
+* @title StakingEscrow
 * @notice Contract holds and locks stakers tokens.
 * Each staker that locks their tokens will receive some compensation
 * @dev |v5.2.2|
